# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Flamingo-Art Open-Api Documentation",
        "description": "Diploma project",
        "contact": {
            "email": "roma.leviczkij@bk.ru"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "2.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api/v1",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/api/chat/get-chat-channel": {
            "get": {
                "tags": [
                    "Chat"
                ],
                "summary": "Get chat",
                "description": "Method that returns chat by user id",
                "operationId": "getChat",
                "parameters": [
                    {
                        "name": "to",
                        "in": "path",
                        "description": "user id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "chat": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/chat/chat-init": {
            "get": {
                "tags": [
                    "Chat"
                ],
                "summary": "Get chat",
                "description": "Method that creating chat in DB",
                "operationId": "chatInit",
                "parameters": [
                    {
                        "name": "to",
                        "in": "path",
                        "description": "from",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "from",
                        "in": "path",
                        "description": "to",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/messages/send-message": {
            "post": {
                "tags": [
                    "Chat"
                ],
                "summary": "Send message",
                "description": "Returns message",
                "operationId": "sendMessage",
                "parameters": [
                    {
                        "name": "to",
                        "in": "path",
                        "description": "from",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "from",
                        "in": "path",
                        "description": "to",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "chat_id",
                        "in": "path",
                        "description": "to",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "message_text",
                        "in": "path",
                        "description": "to",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/messages/get-chat-messages": {
            "get": {
                "tags": [
                    "Chat"
                ],
                "summary": "Get chat-messages",
                "description": "Returns chat messages",
                "operationId": "geChatMessages",
                "parameters": [
                    {
                        "name": "first",
                        "in": "path",
                        "description": "from",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "to",
                        "in": "path",
                        "description": "to",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "chat_id",
                        "in": "path",
                        "description": "to",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "messages": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/messages/check-message": {
            "post": {
                "tags": [
                    "Chat"
                ],
                "summary": "Check message",
                "description": "Method that checks message and return status",
                "operationId": "checkMessage",
                "parameters": [
                    {
                        "name": "message_id",
                        "in": "path",
                        "description": "from",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolen"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
...
    "security": [
        []
    ]
}